Title Screen Update
I added fonts to the app that will allow us to use multiple fonts
through the app. Added a title to the main screen. Also added a video
player to the tile screen, I thought maybe we could add a “What is a
guitar tab?” video. I also added a “Powered by ‘Kivy’” subtitle, when
the kivy image is clicked it opens up the web browser and take the use
to the kivy site.