Fixed issues with the browser
I also moved the video back to the left side and it shouldn’t interfere
with anything (tested browser and the video on my computer both work
great). Also to activate the url you have to click in between the image
and the text. I can’t seem to get it lined up correctly.